set power balance, update unit test

+++ pallets/economy/src/mock.rs
@@ -271,7 +271,7 @@ parameter_types! {
 }
 
 impl orml_nft::Config for Runtime {
-	type ClassId = u64;
+	type ClassId = u32;
 	type TokenId = u64;
 	type ClassData = pallet_nft::NftClassData<Balance>;
 	type TokenData = pallet_nft::NftAssetData<Balance>;
